backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0

# CORS & Filters
django-cors-headers==4.3.1
django-filter==23.5

# Async & Background Jobs
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# Database
psycopg2-binary>=2.9.10

# Payments
stripe==7.6.0

# Auth & Users
django-allauth==0.57.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Media
Pillow

# Admin & Dev Tools
django-import-export==3.3.0
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Docs
drf-yasg==1.21.7

# External Services
firebase-admin==6.3.0
python-telegram-bot==20.6
sentry-sdk==1.38.0

# Activity tracking
django-activity-stream==2.0.0
django-cleanup